projects/poap_cert/pyproject.toml
[tool.poetry]
name = "poap_cert"
version = "0.1.0"
description = "Algorand smart contracts"
authors = ["akhila <akhila.gangisetty001@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
algokit-utils = "^4.0.0"
python-dotenv = "^1.0.0"
py-algorand-sdk = "^2.9.1"
beaker-pyteal = "*"

[tool.poetry.group.dev.dependencies]
algokit-client-generator = "^2.1.0"
puyapy = "*"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

